Fix: Force task execution
This seems to have largely fixed another miss of one or two emails
rarely being missed. It should have been added when switching to the
thread-safe ConcurrentBag<T>.

If the issue persists after further testing, another solution will need
to be researched and a possible refactoring to not use concurrency might
need to be completed.